Add tests for Azure AI agents helpers
