+++ providers/microsoft/azure/tests/unit/microsoft/azure/hooks/test_ai_agents.py
@@ -23,11 +23,6 @@ import pytest
 from azure.core.exceptions import ResourceNotFoundError
 
 from airflow.models import Connection
-from airflow.providers.microsoft.azure._ai_agents import (
-    _get_agent_version,
-    _get_version_status,
-    _serialize_resource,
-)
 from airflow.providers.microsoft.azure.hooks.ai_agents import (
     DEFAULT_REQUEST_TIMEOUT,
     AzureAIAgentsAsyncHook,
@@ -371,48 +366,6 @@ class TestAzureAIAgentsHook:
             hook.invoke_agent_invocations(agent_name=AGENT_NAME, input_data={"message": "hello"})
 
         response.raise_for_status.assert_called_once_with()
-
-    def test_get_version_status_raises_when_status_missing(self):
-        with pytest.raises(ValueError, match="did not include a status"):
-            _get_version_status({})
-
-    def test_get_version_status_normalizes_enum_values(self):
-        class FakeEnum:
-            value = "Active"
-
-        assert _get_version_status({"status": FakeEnum()}) == "active"
-
-    def test_get_agent_version_raises_when_version_missing(self):
-        with pytest.raises(ValueError, match="did not include a version"):
-            _get_agent_version({})
-
-    def test_serialize_resource_uses_as_dict_serializer(self):
-        class SdkModel:
-            def as_dict(self):
-                return {"name": AGENT_NAME, "versions": [{"version": "1"}]}
-
-        assert _serialize_resource(SdkModel()) == {"name": AGENT_NAME, "versions": [{"version": "1"}]}
-
-    def test_serialize_resource_uses_model_dump_serializer(self):
-        class PydanticModel:
-            def model_dump(self):
-                return {"output_text": "hello"}
-
-        assert _serialize_resource(PydanticModel()) == {"output_text": "hello"}
-
-    def test_serialize_resource_rejects_arbitrary_objects(self):
-        class ArbitraryObject:
-            value = "class-value"
-
-            def __init__(self):
-                self.value = "instance-value"
-
-        with pytest.raises(TypeError, match="Cannot serialize.*ArbitraryObject.*for XCom"):
-            _serialize_resource(ArbitraryObject())
-
-    def test_serialize_resource_rejects_non_string_mapping_keys(self):
-        with pytest.raises(TypeError, match="must use string keys"):
-            _serialize_resource({1: "value"})
 
 
 class TestAzureAIAgentsAsyncHook:
